fixup! Settings: Show single-touch wakeup setting when config_performantAuthDefault is true
Not sure why Google assumed that showing this
setting is only useful for devices with FP
sensor built into a side-mounted power button.

The old, desirable behavior was that devices that
supported it could wake up & unlock with a single
touch of the FP sensor.

Removing that from devices with non-UDFPS FP
sensors that don't happen to be built into the
power button is a huge regression.

Change logic to preserve Google's illogical logic
Signed-off-by: ImPrashantt <prashant33968@gmail.com>